requirements.txt
llama-cpp-python 
streamlit